Fix: make sure mask is extended to full size of address
Without that cast, on 64-bit mask was only 32-bit and was cutting down
higher part of pointer, resulting in hardware writing in wrong place
(often over code sections of libraries)

+++ workbench/devs/AHI/Drivers/HDAudio/misc.c
@@ -700,7 +700,7 @@ void *pci_alloc_consistent(size_t size, APTR *NonAlignedAddress, unsigned int bo
     if (address != NULL)
     {
         a = (unsigned long) address;
-        a = (a + boundary - 1) & ~(boundary - 1);
+        a = (a + boundary - 1) & ~((unsigned long)boundary - 1);
         address = (void *) a;
     }
